formatted email and cleaned up code
